split e2e/helpers/env.ts into require-env.ts and test-user.ts; split getVerificationLink out of mailpit.ts into email.ts
