+++ e2e/specs/signup.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from '../fixtures';
 import { dismissCookieBanner } from '../helpers/cookie-banner';
-import { usernamePrefix, emailSuffix, password } from '../helpers/env';
-import { getVerificationLink } from '../helpers/mailpit';
+import { usernamePrefix, emailSuffix, password } from '../helpers/test-user';
+import { getVerificationLink } from '../helpers/email';
 
 test.describe('signup and email verification', () => {
   test('can sign up and verify email via the emailed link', async ({
